refactor(workspace): give fractional-index its own subpath, drop the attach-timeline export
computeMidpoint and generateInitialOrders are generic ordering math, but they
sat inside the timeline folder and were the only symbols anyone imported from
the @epicenter/workspace/document/attach-timeline subpath. @epicenter/filesystem
even called that subpath the 'canonical location' for ordering helpers.

Move the file to src/document/fractional-index.ts and ship it from a
@epicenter/workspace/document/fractional-index subpath. Remove the
attach-timeline subpath: attachTimeline is exported from the package root, and
nothing imported the sheet serialization or entry types through it. Repoint
filesystem to the new subpath.
