+++ packages/workspace/src/document/fractional-index.ts
